Fix mail_client_service structure and restore coverage to 87.96%
- Moved _impl.py and generated/ from nested src/ to correct location
- Fixed test fixture to properly mock mail_client_api.Client
- Moved test_service_client.py to src/mail_client_service/tests/
- Updated test assertions to handle both mock clients
- Excluded auto-generated code from coverage reporting
- All 86 tests passing with 87.96% coverage
